# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "重生之凡人修仙传 API",
    "description": "修仙文字游戏后端接口文档，基于《凡人修仙传》世界观。包含玩家注册登录、角色管理、境界突破、地图探索、闭关修炼、GM管理、系统通知、WebSocket实时推送、玩家数据实时同步等完整功能。",
    "version": "2.2.0",
    "contact": {
      "name": "Shell",
      "email": "1592363624@qq.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发环境"
    },
    {
      "url": "https://xx.52shell.ltd",
      "description": "生产环境"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证"
      }
    },
    "schemas": {
...
      "SystemNotification": {
        "type": "object",
        "description": "系统通知",
        "properties": {
          "id": {
            "type": "integer",
            "description": "通知ID"
          },
          "type": {
            "type": "string",
            "enum": ["breakthrough", "death", "achievement", "event", "announcement", "warning", "milestone"],
            "description": "通知类型：突破、死亡、成就、事件、公告、警告、里程碑"
          },
          "title": {
            "type": "string",
            "description": "通知标题"
          },
          "content": {
            "type": "string",
            "description": "通知内容"
          },
          "priority": {
            "type": "string",
            "enum": ["low", "normal", "high", "critical"],
            "description": "优先级"
          },
          "actorNickname": {
            "type": "string",
            "description": "触发事件玩家昵称"
          },
          "isRead": {
            "type": "boolean",
            "description": "是否已读"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/notifications": {
      "get": {
        "summary": "获取通知列表",
        "description": "获取当前玩家的通知列表，支持分页和筛选",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "页码"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "每页数量，最大100"
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["breakthrough", "death", "achievement", "event", "announcement", "warning", "milestone"]
            },
            "description": "通知类型筛选"
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "仅返回未读通知"
          },
          {
            "name": "includeGlobal",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "是否包含全服通知"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "total": {
                      "type": "integer",
                      "description": "总数量"
                    },
                    "page": {
                      "type": "integer",
                      "description": "当前页"
                    },
                    "limit": {
                      "type": "integer",
                      "description": "每页数量"
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "总页数"
                    },
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SystemNotification"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/unread-count": {
      "get": {
        "summary": "获取未读通知数量",
        "description": "获取当前玩家未读通知的数量",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "count": {
                      "type": "integer",
                      "description": "未读通知数量"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/global": {
      "get": {
        "summary": "获取全服重要通知",
        "description": "获取全服广播的重要通知，用于新玩家登录时显示",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "获取数量"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SystemNotification"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "post": {
        "summary": "标记通知为已读",
        "description": "将指定通知标记为已读状态",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "通知ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/read-all": {
      "post": {
        "summary": "标记所有通知为已读",
        "description": "将当前玩家的所有通知标记为已读状态",
        "tags": ["Notification"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/announcement": {
      "post": {
        "summary": "发送全服公告",
        "description": "向全服玩家发送公告通知（需要管理员权限）",
        "tags": ["Notification", "GM"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "公告标题"
                  },
                  "content": {
                    "type": "string",
                    "description": "公告内容"
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["low", "normal", "high", "critical"],
                    "default": "high",
                    "description": "优先级"
                  }
                }
              },
              "examples": {
                "normal": {
                  "value": {
                    "title": "服务器维护通知",
                    "content": "服务器将于今晚22:00-24:00进行维护，届时将无法登录。",
                    "priority": "high"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "公告已发送"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "example": {
                  "code": 403,
                  "message": "权限不足"
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/event": {
      "post": {
        "summary": "发送事件通知",
        "description": "发送玩家事件通知，如突破、死亡等（需要管理员权限）",
        "tags": ["Notification", "GM"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["eventName", "content"],
                "properties": {
                  "eventName": {
                    "type": "string",
                    "description": "事件名称"
                  },
                  "content": {
                    "type": "string",
                    "description": "事件内容"
                  },
                  "priority": {
                    "type": "string",
                    "enum": ["low", "normal", "high", "critical"],
                    "default": "high",
                    "description": "优先级"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "发送成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "事件通知已发送"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/notifications/{id}": {
      "delete": {
        "summary": "删除通知 (GM)",
        "description": "删除指定通知（需要管理员权限）",
        "tags": ["Admin", "GM"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "通知ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "通知删除成功"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "通知不存在",
            "content": {
              "application/json": {
                "example": {
                  "code": 404,
                  "message": "通知不存在"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "example": {
                  "code": 403,
                  "message": "权限不足"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}